rngstream changes
Minimal changes to rngstream class, just removal of unsigned long (varies between compiler)

Then moved rngstream to a vector, with a bit of fiddling to make everything happy with that.

+++ src/RngStream.cpp
@@ -18,7 +18,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with RngStreams.  If not, see <http://www.gnu.org/licenses/>.
+ * aint32_t with RngStreams.  If not, see <http://www.gnu.org/licenses/>.
  *
  * We would appreciate that 
  * if you use this software for work leading to publications, 
@@ -29,7 +29,7 @@
  * Operations Research, 47, 1 (1999), 159--164.
 
  * P. L'Ecuyer, R. Simard, E. J. Chen, and W. D. Kelton, 
- * ``An Objected-Oriented Random-Number Package with Many Long Streams and Substreams'', 
+ * ``An Objected-Oriented Random-Number Package with Many int32_t Streams and Substreams'', 
  * Operations Research, 50, 6 (2002), 1073--1075
  *
 \***********************************************************************/
@@ -112,18 +112,18 @@ const double A2p127[3][3] = {
 double MultModM (double a, double s, double c, double m)
 {
     double v;
-    long a1;
+    int32_t a1;
 
     v = a * s + c;
 
     if (v >= two53 || v <= -two53) {
-        a1 = static_cast<long> (a / two17);    a -= a1 * two17;
+        a1 = static_cast<int32_t> (a / two17);    a -= a1 * two17;
         v  = a1 * s;
-        a1 = static_cast<long> (v / m);     v -= a1 * m;
+        a1 = static_cast<int32_t> (v / m);     v -= a1 * m;
         v = v * two17 + a * s + c;
     }
 
-    a1 = static_cast<long> (v / m);
+    a1 = static_cast<int32_t> (v / m);
     /* in case v < 0)*/
     if ((v -= a1 * m) < 0.0) return v += m;   else return v;
 }
@@ -175,7 +175,7 @@ void MatMatModM (const double A[3][3], const double B[3][3],
 //-------------------------------------------------------------------------
 // Compute the matrix B = (A^(2^e) Mod m);  works also if A = B. 
 //
-void MatTwoPowModM (const double A[3][3], double B[3][3], double m, long e)
+void MatTwoPowModM (const double A[3][3], double B[3][3], double m, int32_t e)
 {
    int i, j;
 
@@ -194,7 +194,7 @@ void MatTwoPowModM (const double A[3][3], double B[3][3], double m, long e)
 //-------------------------------------------------------------------------
 // Compute the matrix B = (A^n Mod m);  works even if A = B.
 //
-void MatPowModM (const double A[3][3], double B[3][3], double m, long n)
+void MatPowModM (const double A[3][3], double B[3][3], double m, int32_t n)
 {
     int i, j;
     double W[3][3];
@@ -221,7 +221,7 @@ void MatPowModM (const double A[3][3], double B[3][3], double m, long n)
 // Check that the seeds are legitimate values. Returns 0 if legal seeds,
 // -1 otherwise.
 //
-int CheckSeed (const unsigned long seed[6])
+int CheckSeed (const uint32_t seed[6])
 {
     int i;
 
@@ -265,19 +265,19 @@ int CheckSeed (const unsigned long seed[6])
 //
 double RngStream::U01 ()
 {
-    long k;
+    int32_t k;
     double p1, p2, u;
 
     /* Component 1 */
     p1 = a12 * Cg[1] - a13n * Cg[0];
-    k = static_cast<long> (p1 / m1);
+    k = static_cast<int32_t> (p1 / m1);
     p1 -= k * m1;
     if (p1 < 0.0) p1 += m1;
     Cg[0] = Cg[1]; Cg[1] = Cg[2]; Cg[2] = p1;
 
     /* Component 2 */
     p2 = a21 * Cg[5] - a23n * Cg[3];
-    k = static_cast<long> (p2 / m2);
+    k = static_cast<int32_t> (p2 / m2);
     p2 -= k * m2;
     if (p2 < 0.0) p2 += m2;
     Cg[3] = Cg[4]; Cg[4] = Cg[5]; Cg[5] = p2;
@@ -378,7 +378,7 @@ void RngStream::ResetNextSubstream ()
 
 
 //-------------------------------------------------------------------------
-bool RngStream::SetPackageSeed (const unsigned long seed[6])
+bool RngStream::SetPackageSeed (const uint32_t seed[6])
 {
    if (CheckSeed (seed))
       return false;                   // FAILURE     
@@ -389,7 +389,7 @@ bool RngStream::SetPackageSeed (const unsigned long seed[6])
 
 
 //-------------------------------------------------------------------------
-bool RngStream::SetSeed (const unsigned long seed[6])
+bool RngStream::SetSeed (const uint32_t seed[6])
 {
    if (CheckSeed (seed))
       return false;                   // FAILURE     
@@ -405,7 +405,7 @@ bool RngStream::SetSeed (const unsigned long seed[6])
 // if e = 0, let n = c.
 // Jump n steps forward if n > 0, backwards if n < 0.
 //
-void RngStream::AdvanceState (long e, long c)
+void RngStream::AdvanceState (int32_t e, int32_t c)
 {
     double B1[3][3], C1[3][3], B2[3][3], C2[3][3];
 
@@ -436,10 +436,10 @@ void RngStream::AdvanceState (long e, long c)
 
 
 //-------------------------------------------------------------------------
-void RngStream::GetState (unsigned long seed[6]) const
+void RngStream::GetState (uint32_t seed[6]) const
 {
    for (int i = 0; i < 6; ++i)
-      seed[i] = static_cast<unsigned long> (Cg[i]);
+      seed[i] = static_cast<uint32_t> (Cg[i]);
 }
 
 
@@ -452,9 +452,9 @@ void RngStream::WriteState () const
     cout << ":\n   Cg = { ";
 
     for (int i = 0; i < 5; i++) {
-        cout << static_cast<unsigned long> (Cg [i]) << ", ";
+        cout << static_cast<uint32_t> (Cg [i]) << ", ";
     }
-    cout << static_cast<unsigned long> (Cg [5]) << " }\n\n";
+    cout << static_cast<uint32_t> (Cg [5]) << " }\n\n";
 }
 
 
@@ -471,21 +471,21 @@ void RngStream::WriteStateFull () const
 
     cout << "   Ig = { ";
     for (i = 0; i < 5; i++) {
-        cout << static_cast<unsigned long> (Ig [i]) << ", ";
+        cout << static_cast<uint32_t> (Ig [i]) << ", ";
     }
-    cout << static_cast<unsigned long> (Ig [5]) << " }\n";
+    cout << static_cast<uint32_t> (Ig [5]) << " }\n";
 
     cout << "   Bg = { ";
     for (i = 0; i < 5; i++) {
-        cout << static_cast<unsigned long> (Bg [i]) << ", ";
+        cout << static_cast<uint32_t> (Bg [i]) << ", ";
     }
-    cout << static_cast<unsigned long> (Bg [5]) << " }\n";
+    cout << static_cast<uint32_t> (Bg [5]) << " }\n";
 
     cout << "   Cg = { ";
     for (i = 0; i < 5; i++) {
-        cout << static_cast<unsigned long> (Cg [i]) << ", ";
+        cout << static_cast<uint32_t> (Cg [i]) << ", ";
     }
-    cout << static_cast<unsigned long> (Cg [5]) << " }\n\n";
+    cout << static_cast<uint32_t> (Cg [5]) << " }\n\n";
 }
